refactor(extension): move `matchQueryResult` to `query` folder
